requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Django==3.2.8
django-js-asset==1.2.2
Flask==2.0.2
Flask-Script==2.0.6
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
pycryptodome==3.12.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
Werkzeug==2.0.2
